final push

+++ C06_Module/ex02/main.cpp
@@ -7,7 +7,7 @@
 #include "C.hpp"
 
 int main() {
-    srand(static_cast<unsigned int>(time(0)));
+    srand((time(0)));
 
     Base* instance = generate();
